Version_finale_avec_erreur

+++ reseau/diffusion.h
@@ -1,3 +1,9 @@
+/**
+ * @file diffusion.h
+ * @brief Structures du protocole UDP et gestion de la fiabilité.
+ */
+
+
 #ifndef DIFFUSION__h
 #define DIFFUSION__h
 
@@ -15,18 +21,21 @@
 
 
 // Cette directive magique empêche le compilateur C de rajouter des octets 
-// vides (padding) au milieu de ta structure. C'est vital pour le réseau !
+// vides (padding) au milieu de la structure. 
 #pragma pack(push, 1)
 
 
 
-// Structure qui va envelopper le paquet a envoyer
+/**
+ * @struct EnteteUDP
+ * @brief En-tête personnalisé pour les paquets réseau du jeu.
+ */
 typedef struct
 {
-    uint32_t id_expediteur;     // L'identifiant du joueur
-    uint8_t type_message;      // Le type du message 
-    uint32_t num_sequence;     // Le numero de suivie 
-    uint16_t taille_payload;   // La taille des donnees JSON, pour lire le json (2 octets)
+    uint32_t id_expediteur;     /**< ID du joueur émetteur. */
+    uint8_t type_message;       /**< Type : 0:Jeu, 1:ACK, 2:Ping, etc.. */
+    uint32_t num_sequence;      /**< ID du paquet pour acquittement (htonl). */
+    uint16_t taille_payload;    /**< Taille du JSON (htons). */
 } EnteteUDP;
 
 
@@ -34,17 +43,20 @@ typedef struct
 
 
 
-// Structure pour ta "Salle des Machines" (Ne part pas sur le réseau)
+/**
+ * @struct NoeudAttente
+ * @brief Élément d'une liste chaînée pour la retransmission des messages.
+ */
 typedef struct NoeudAttente {
-    EnteteUDP entete;               // La copie de l'en-tête envoyé
-    struct sockaddr_in dest;        // Destination du paquet
-    char payload[10049];             // La copie de ton JSON 
-    long temps_envoi;               // Le chronomètre (pour savoir quand renvoyer)
-    struct NoeudAttente *suivant;   // Le pointeur vers le colis suivant dans la liste !
+    EnteteUDP entete;                /**< Copie de l'en-tête envoyé. */
+    struct sockaddr_in dest;         /**< Destination du message. */
+    char payload[10049];             /**< Copie du contenu JSON. */
+    long temps_envoi;                /**< Heure du dernier envoi (ms). */
+    struct NoeudAttente *suivant;    /**< Pointeur vers le message suivant. */
 } NoeudAttente;
 
 
-/*------------------------FONCTIONS--------------------------------- */
+// Prototypes
 extern int diffusion_message_sens1(const char *donnee_json, int mon_socket_udp, uint8_t type_msg);
 extern char *diffusion_message_sens2(int reseau_fd);
 extern void verifier_retransmissions(int mon_socket_udp);
@@ -52,11 +64,7 @@ extern long get_time();
 extern void supprimer_de_la_file(uint32_t seq_a_supprimer, struct sockaddr_in expediteur);
 extern void message_systeme(int mon_socket_udp, uint8_t type_msg, uint32_t num_seq, struct sockaddr_in dest);
 extern void nettoyer_file_joueur_parti(struct sockaddr_in joueur_parti);
-
-extern void set_mon_id(uint8_t id);
-
-
-
+extern void set_mon_id(uint32_t id);
 
 
 #endif
